+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## `ipmi-rs-core`
 * Add support for the `SetUserAccess` and `GetUserAccess` commands.
+* Add support for the `SetUserName` and `GetUserName` commands.
 
 # [0.5.0](https://github.com/datdenkikniet/ipmi-rs/tree/v0.5.0)